Fix warning in NPC.cpp: use size_t for variable.

+++ src/game/NPC.cpp
@@ -1444,7 +1444,7 @@ long IsNearSelection(EERIE_3DOBJ * obj, long vert, long tw)
 
 	if (vert < 0) return -1;
 
-	for (long i = 0; i < obj->selections[tw].selected.size(); i++)
+	for (size_t i = 0; i < obj->selections[tw].selected.size(); i++)
 	{
 		float d = TRUEEEDistance3D(&obj->vertexlist[obj->selections[tw].selected[i]].v,
 		                           &obj->vertexlist[vert].v);
